Poke4Outlook.exe
A script with the same purpose in a directly executable format. It will provide an error message if Outlook is not installed on your system.

A script with the same purpose in a directly executable format. It will give an error if you do not have outlook on your system.